Moved RawFileTransmissionData to schema file
