+++ src/nomad_measurements/transmission/parser.py
@@ -17,32 +17,17 @@
 #
 from typing import TYPE_CHECKING
 
-from nomad.datamodel.data import (
-    EntryData,
-)
-from nomad.datamodel.metainfo.annotations import (
-    ELNAnnotation,
-)
-from nomad.metainfo import (
-    Quantity,
-)
 from nomad.parsing import MatchingParser
 
-from nomad_measurements.transmission.schema import ELNTransmission
+from nomad_measurements.transmission.schema import (
+    ELNTransmission,
+    RawFileTransmissionData,
+)
 from nomad_measurements.utils import create_archive
 
 if TYPE_CHECKING:
     from nomad.datamodel.datamodel import (
         EntryArchive,
-    )
-
-
-class RawFileTransmissionData(EntryData):
-    measurement = Quantity(
-        type=ELNTransmission,
-        a_eln=ELNAnnotation(
-            component='ReferenceEditQuantity',
-        ),
     )
