+++ tests/test_utils.py
@@ -22,7 +22,6 @@ from nomad.datamodel.metainfo.basesections import (
     PureSubstanceComponent,
     PureSubstanceSection,
 )
-
 from nomad_measurements.utils import (
     merge_sections,
 )
